basic animation function

+++ src/main/java/top/whitecola/animationlib/functions/IAnimationFunction.java
@@ -1,7 +1,7 @@
 package top.whitecola.animationlib.functions;
 
 public interface IAnimationFunction {
-    float get(float pastTime);
+    float get(long pastTime);
 
 
 }
